+++ .changeset/friendly-turtles-behave.md
@@ -2,4 +2,5 @@
 "@salt-ds/lab": minor
 ---
 
+Moved `useFormFieldPropsNext` and `FormFieldContextNext` to `./form-field-context-next`
 Move `a11yValueAriaProps`, `A11yValueProps` to `FormFieldContextNext`
