+++ packages/lab/src/form-field-context/index.ts
@@ -1,4 +1,2 @@
 export * from "./FormFieldContext";
-export * from "./FormFieldContextNext";
 export * from "./useFormFieldProps";
-export * from "./useFormFieldPropsNext";
